Issue #107 - Reorganize the app with routes
